refactor(Filters): refactor filters components and update styles
Moved Filters, FiltersList, and FiltersListCheckbox logic to dedicated files for better modularity. Updated imports across related components and tests. Added ExcursionsPage.module.css for mobile filter bar styling and moved filter badge/button styles out of Layout.module.css. Adjusted NavigationBar z-index for proper stacking with sticky elements.
