+++ src/components/Layout/Layout.module.css
@@ -37,52 +37,3 @@ html {
 	flex-direction: column;
 	flex-grow: 1;
 }
-
-.filterBadge {
-	background-color: var(--login-btn-bg);
-	color: var(--login-btn-text);
-	font-weight: 600;
-	font-size: 0.8em;
-	padding: 0.5em 0.8em;
-	border: 1px solid transparent; /* Se añade un borde para que la transición del color de borde funcione */
-}
-
-/*
- * En modo oscuro, se ajustan los colores del badge para que sean consistentes con el texto del botón (claro sobre fondo oscuro).
- * Se usa :global() para apuntar a la clase del tema desde el módulo CSS.
-*/
-:global(html.dark) .filterBadge {
-	background-color: var(--heading-color);
-	color: var(
-		--login-btn-text
-	); /* Se usa el mismo color de texto que el botón de login para asegurar contraste */
-}
-
-/*
- * Cuando el botón de filtros está activo (presionado), se invierten los colores del badge
- * para que coincida con el comportamiento del texto del botón.
- * Se usa :global() para apuntar a la clase global del botón desde el módulo CSS.
-*/
-:global(.filtersToggleButton:active) .filterBadge {
-	/* Se cambian los colores para que el badge reaccione igual que el botón. */
-	background-color: var(--filters-toggle-btn-bg-hover);
-	color: var(--filters-toggle-btn-color-hover);
-	border-color: var(--filters-toggle-btn-border-hover);
-}
-
-/* --- Offcanvas (Filtros) --- */
-/* Botón para mostrar filtros en breakpoints pequeños */
-/* Se usa .btn para aumentar la especificidad y sobreescribir los estilos por defecto de Bootstrap */
-.filtersToggleButton {
-	color: var(--filters-toggle-btn-color);
-	border: 1px solid var(--filters-toggle-btn-border);
-	background-color: transparent;
-	transition: none;
-}
-
-.filtersToggleButton:hover,
-.filtersToggleButton:active {
-	color: var(--filters-toggle-btn-color-hover);
-	background-color: var(--filters-toggle-btn-bg-hover);
-	border-color: var(--filters-toggle-btn-border-hover);
-}
